Fix Deezer events not merging with Ticketmaster on upcoming page.
Deezer liveEvents omits venue state while TM includes it, so dedupe keys split and DZ sources sat on duplicate rows without lat/lon that the radius filter excluded. Drop region from named-venue fingerprints and re-coalesce existing rows on upgrade.
